chore: extract the common "interface RTT and GStreamer" code into a Common task
This class is meant to provide the services needed to build
other types of components, as e.g. a refactoring of the webrtc
streamer/receiver.
