+++ tasks/Helpers.hpp
@@ -63,9 +63,10 @@ namespace gstreamer {
         }
     };
 
-    static GstVideoFormat frameModeToGSTFormat(base::samples::frame::frame_mode_t format)
+    inline GstVideoFormat rawModeToGSTVideoFormat(
+        base::samples::frame::frame_mode_t frame_mode)
     {
-        switch (format) {
+        switch (frame_mode) {
             case base::samples::frame::MODE_RGB:
                 return GST_VIDEO_FORMAT_RGB;
             case base::samples::frame::MODE_BGR:
@@ -77,7 +78,79 @@ namespace gstreamer {
             default:
                 // Should not happen, the component validates the frame mode
                 // against the accepted modes
-                throw std::runtime_error("unsupported frame format received");
+                throw std::runtime_error("unsupported raw format received");
+        }
+    }
+
+    inline GstCaps* rawModeToGSTCaps(base::samples::frame::frame_mode_t frame_mode)
+    {
+        auto format = rawModeToGSTVideoFormat(frame_mode);
+        GstCaps* caps = gst_caps_new_simple("video/x-raw",
+            "format",
+            G_TYPE_STRING,
+            gst_video_format_to_string(format),
+            NULL);
+        if (!caps) {
+            throw std::runtime_error("failed to generate caps");
+        }
+
+        return caps;
+    }
+
+    inline std::string bayerModeToGSTCapsFormat(
+        base::samples::frame::frame_mode_t frame_mode)
+    {
+        switch (frame_mode) {
+            case base::samples::frame::MODE_BAYER_BGGR:
+                return "bggr";
+            case base::samples::frame::MODE_BAYER_GBRG:
+                return "gbrg";
+            case base::samples::frame::MODE_BAYER_GRBG:
+                return "grbg";
+            case base::samples::frame::MODE_BAYER_RGGB:
+                return "rggb";
+            default:
+                // Should not happen, the component validates the frame mode
+                // against the accepted modes
+                throw std::runtime_error("unsupported bayer format received");
+        }
+    }
+
+    inline GstCaps* bayerModeToGSTCaps(base::samples::frame::frame_mode_t frame_mode)
+    {
+        auto format = bayerModeToGSTCapsFormat(frame_mode);
+        GstCaps* caps = gst_caps_new_simple("video/x-bayer",
+            "format",
+            G_TYPE_STRING,
+            format.c_str(),
+            NULL);
+        if (!caps) {
+            throw std::runtime_error("failed to generate caps");
+        }
+
+        return caps;
+    }
+
+    inline bool isFrameModeBayer(base::samples::frame::frame_mode_t frame_mode)
+    {
+        switch (frame_mode) {
+            case base::samples::frame::MODE_BAYER_BGGR:
+            case base::samples::frame::MODE_BAYER_GBRG:
+            case base::samples::frame::MODE_BAYER_GRBG:
+            case base::samples::frame::MODE_BAYER_RGGB:
+                return true;
+            default:
+                return false;
+        }
+    }
+
+    inline GstCaps* frameModeToGSTCaps(base::samples::frame::frame_mode_t frame_mode)
+    {
+        if (isFrameModeBayer(frame_mode)) {
+            return bayerModeToGSTCaps(frame_mode);
+        }
+        else {
+            return rawModeToGSTCaps(frame_mode);
         }
     }
 }
